+++ drishti/viewinformation.cpp
@@ -454,7 +454,7 @@ ViewInformation::save(fstream &fout)
   keyword = "rotation";
   fout.write((char*)(keyword.toLatin1().data()), keyword.length()+1);  
   Vec axis;
-  float angle;
+  qreal angle;
   m_rotation.getAxisAngle(axis, angle);
   f[0] = axis.x;
   f[1] = axis.y;
